API-Scripts/requirements.txt
requests>=2.24.0
python-dotenv>=1.0.1